Fix game of life
Make it a normal example. Remove definitions from Agents.jl
