+++ src/Agents.jl
@@ -22,6 +22,5 @@ include("simulations/sample.jl")
 
 # Auxilary
 include("CA1D.jl")
-include("CA2D.jl")
 
 end # module
